==> 2_processing/1-preparacion-datos_nse.Rmd <==
---
title: "Código de preparación de datos"
author: "Cristóbal Ortiz"
output: html_document
---

```{r set-markdown,include=FALSE}
knitr::opts_chunk$set(warning = FALSE,message = FALSE, cache = TRUE)
knitr::opts_knit$set(root.dir = "../")
options(scipen=999)
options(knitr.kable.NA = '')
```

```{r load-library, message=FALSE, warning=FALSE}
library(tidyverse) #herramientas para manipulación y visualización de datos (incluye ggplot2)
library(openxlsx) #importar/exportar tablas de excel
library(sjmisc) #manejo e inspección de variables
library(summarytools) # datos resumidos
library(sjlabelled) #manejo de etiquetas de variables
library(scales) #reescalamiento y estandarización de variables
library(sf) #manejo de base de datos espaciales
library(ismtchile) #indice socio-material territorial de Chile
library(chilemapas) #coberturas y geometrías de mapas en Chile
```

```{r import-dataset}
#getwd()
remove(list = ls())

#load datasets
load(url("https://dataverse.harvard.edu/api/access/datafile/7245115")) #elsoc wide 2016-2022
manzanas_2017_elsoc <- haven::read_dta("../1_input/data/original/elsoc-manzanas/mzn-elsoc-2012-2017/manzanas2017_elsoc_2016_2019.dta") %>%
  dplyr::select(idencuesta, cod_zona = geocodigo_ok)
ismt_data <- read.csv("../1_input/data/original/ocuc-ismt/ismt_rm_2017.csv") %>% #ismt calculado por OCUC
  dplyr::select(cod_zona = zona, ismt_rank = ismtpn) #selección y renombre de variables
```

```{r join-datasets}

## JOIN: ELSOC & ZONAS
manzanas_2017_elsoc$idencuesta <- as.numeric(manzanas_2017_elsoc$idencuesta) #cambio clase para hacer el siguiente join
elsoc_a <- dplyr::left_join(x = elsoc_wide_2016_2022.2, y = manzanas_2017_elsoc, by = "idencuesta")

## JOIN: ELSOC-ZONAS & ISMT
elsoc_a$cod_zona <- as.numeric(elsoc_a$cod_zona) #cambio clase para hacer el siguiente join
elsoc_b <- dplyr::left_join(x = elsoc_a, y = ismt_data, by = "cod_zona")
```

```{r clean-dataset}
elsoc_ams <- elsoc_b %>% 
  dplyr::select(idencuesta, #id individual
                estrato_disenno, estrato_w06, comuna_w06, comuna_cod_w06, cod_zona, #id geográficos
                starts_with("t02_"), #pertenencia barrial 
                starts_with("t03_"), #sociabilidad barrial
                starts_with("t06_02") ,starts_with("t06_05"), starts_with("t06_06"), starts_with("t06_07"),  #satisfacción barrial
                starts_with("t08_"), starts_with("t10_"), starts_with("t04_06_"), # reputación | seguridad | conflicto barrial 
                ismt_rank, #desigualdad socioespacial
                m0_sexo_w06, m0_edad_w06, #sexo | edad
                starts_with("m01_"), starts_with("d01_01_"), # escolaridad | estatus subjetivo 
                starts_with("m29_"), #ingreso familiar monto
                starts_with("m34_03_")) %>% # tiempo de residencia
  filter(estrato_w06 == 1 &  #solo Gran Santiago en la medición 2022
         comuna_cod_w06 < 13605 & # excluye comunas rurales del RM
         comuna_cod_w06 > 11000 & # excluye comunas fuera de la RM
         !comuna_cod_w06 %in% c(13602, 13603)) # elimina El Monte e Isla de Maipo

#convertimos no sabe/no responde/no disponible/sin dato/no aplica a NA
elsoc_ams[elsoc_ams == -888 | elsoc_ams == -999 | elsoc_ams == 994 | elsoc_ams == 995 | elsoc_ams == 996 | elsoc_ams == -666 | elsoc_ams == -777] <- NA # varias categoría a NA

#view(dfSummary(elsoc_ams, plain.ascii  = FALSE, style = "grid", graph.magnif = 0.75, valid.col = FALSE))

#IMPORTANTE LEER: luego de esta limpieza de datos, se concluye que los códigos de zonas y comunas de ELSOC, difieren en aprox 35 casos respecto a las zonas y comunas de la data que contiene las manzanas reales. Ante la imposibilidad de verificar cuál código es el adecuado, optamos por guiarnos por los códigos de ELSOC (y especificamente la ola 2022). Es decir, la base de datos final contiene aquellos que cumples dos condiciones: en la medición original (2026) vivían en el Gran Santiago; ii) en la última medición (2022) siguen viviendo en el Gran Santiago.
```

```{r imput-intertemporal}
elsoc_ams  <- elsoc_ams %>%
  mutate(across(starts_with("m29_"), ~if_else(. %in% 0, NA_real_, .))) %>% #convierto ingreso = 0 a NA
  mutate(#INGRESO FAMILIAR
         ingr_prom = rowSums(select(., starts_with("m29_")), na.rm = TRUE) / rowSums(!is.na(select(., starts_with("m29_")))),
         #BATERÍA DE PERTENENCIA BARRIAL
         t02_01 = coalesce(t02_01_w06, t02_01_w04, t02_01_w03, t02_01_w02, t02_01_w01), 
         t02_02 = coalesce(t02_02_w06, t02_02_w04, t02_02_w03, t02_02_w02, t02_02_w01),
         t02_03 = coalesce(t02_03_w06, t02_03_w04, t02_03_w03, t02_03_w02, t02_03_w01),
         t02_04 = coalesce(t02_04_w06, t02_04_w04, t02_04_w03, t02_04_w02, t02_04_w01),
         #BATERÍA DE SOCIABILIDAD BARRIAL
         t03_01 = coalesce(t03_01_w06, t03_01_w04, t03_01_w03, t03_01_w02, t03_01_w01), 
         t03_02 = coalesce(t03_02_w06, t03_02_w04, t03_02_w03, t03_02_w02, t03_02_w01),
         t03_03 = coalesce(t03_03_w06, t03_03_w04, t03_03_w03, t03_03_w02, t03_03_w01),
         t03_04 = coalesce(t03_04_w06, t03_04_w04, t03_04_w03, t03_04_w02, t03_04_w01),
         #BATERÍA DE SATISFACCIÓN BARRIAL
         t06_02 = coalesce(t06_02_w06, t06_02_w04, t06_02_w03, t06_02_w02, t06_02_w01),
         t06_05 = coalesce(t06_05_w06, t06_05_w04, t06_05_w03, t06_05_w02, t06_05_w01),
         t06_06 = coalesce(t06_06_w06, t06_06_w04, t06_06_w03, t06_06_w02, t06_06_w01),
         t06_07 = coalesce(t06_07_w06, t06_07_w04, t06_07_w03, t06_07_w02, t06_07_w01),
         #REPUTACIÓN BARRIAL,
         t08 = coalesce(t08_w06, t08_w05, t08_w04, t08_w03, t08_w02, t08_w01),
         #SEGURIDAD BARRIAL
         t10 = coalesce(t10_w06, t10_w05, t10_w04, t10_w03, t10_w02, t10_w01),
         #CONFLICTO INTERGRUPAL,
         t04_06 = coalesce(t04_06_w04, t04_06_w03, t04_06_w02, t04_06_w01),
         #EDUCACIÓN
         m01 = coalesce(m01_w06, m01_w05, m01_w04, m01_w03, m01_w02, m01_w01),
         #ESTATUS SUBJETIVO
         d01_01 = coalesce(d01_01_w06, d01_01_w05, d01_01_w04, d01_01_w03, d01_01_w02, d01_01_w01),
         #TIEMPO DE RESIDENCIA
         m34_03 = coalesce(m34_03_w03, m34_03_w01))
         
```

```{r manipulate-variables}
elsoc_ams <- elsoc_ams %>% 
  #remove_all_labels() %>%
  #-----Recode Y: cohesión barrial
  mutate(spbi = rowMeans(dplyr::select(., t02_01, t02_02, t02_03, t02_04), na.rm = TRUE), #promedio apego al barrio
         soci = rowMeans(dplyr::select(., t03_01, t03_02, t03_03, t03_04), na.rm = TRUE)) %>%  #promedio sociabilidad barrial
  mutate(spbf = factor(cut(spbi, breaks = c(1, 3, 4, 5), 
                           labels = c("Baja", "Media", "Alta"), levels = 1:3)),
         socf = factor(cut(soci, breaks = c(1, 3, 4, 5), 
                           labels = c("Baja", "Media", "Alta"), levels = 1:3)),
  #-----Recode M: mecanismos subjetivos 
         segu = factor(car::recode(t10, "1:2=1;3=2;4:5=3"),
                       labels = c("Baja",  "Media", "Alta"), levels = 1:3),
         segui = t10,
         sacci = rowMeans(dplyr::select(., t06_02, t06_05, t06_06, t06_07), na.rm = TRUE),
         repb = factor(car::recode(t08, "1:2=1;3=2;4:5=3"),
                       labels = c("Negativa", "Neutra", "Positiva"), levels = 1:3),
         repbi = t08,
         desgr = t04_06, 
         desea = ((t04_06 - max(t04_06_w04, na.rm = T)) * -1) + min(t04_06, na.rm = T)) %>% 
  #-----Recode C: controles sociodemográficos
  mutate(edad = factor(car::recode(m0_edad_w06, "18:29=1; 30:49=2; 50:64=3; 65:150=4"),
                       labels = c('18-29', '30-49', '50-64', '65 o más'), levels = 1:4),
         edadi = m0_edad_w06,
         educ = factor(car::recode(m01,"1:3=1; 4:5=2; 6:7=3; 8:10=4"),
                       labels = c("Básica", "Media", "Técnica", "Universitaria"), levels = 1:4),
         educi = m01,
         essu = factor(car::recode(d01_01, "c(6,7,8,9,10)=1; c(4,5)=2; c(0,1,2,3)=3"),
                      labels = c("Alto", "Medio", "Bajo"), levels = 1:3),
         essui = d01_01,
         time = m34_03) %>% 
  #-----Etiqueta variables
  var_labels(spbi = "Pertenencia barrial",
             spbf = "Pertenencia barrial",
             soci = "Sociabilidad barrial",
             socf = "Sociabilidad barrial",
             segu = 'Sentimiento seguridad',
             segui = 'Sentimiento seguridad',
             sacci = 'Satisfacción residencial',
             repb = 'Reputación percibida',
             repbi = 'Reputación percibida',
             desgr = 'Residentes desagradables',
             edad = 'Tramo etario',
             educ = 'Nivel educacional',
             essu = 'Estatus social subjetivo',
             essui = 'Estatus social subjetivo',
             edadi = 'Edad residente',
             educi = "Nivel educacional",
             time = 'Tiempo de residencia',
             ismt_rank = "NSE barrio")

## Estandarización de variables
#elsoc_ams <- elsoc_ams %>% mutate_at(vars(spbi, soci, segui, repbi, desgr, essui, edadi, educi, time, ismt_rank), scale)

view(dfSummary(elsoc_ams, plain.ascii  = FALSE, style = "grid", graph.magnif = 0.75, valid.col = FALSE))

```

```{r join-geometrías}
#CARGAMOS GEOMETRÍA DE LAS ZONAS DEL GRAN SANTIAGO
zonas_ams <- mapa_zonas %>% 
  filter(codigo_region == 13) %>% 
  rename("cod_zona" = "geocodigo") %>% 
  dplyr::select(codigo_provincia, codigo_comuna, cod_zona, geometry) %>% 
  filter(codigo_provincia %in% c(131, 132, 134, 136), #PROVINCIAS DEL GS EXCEPTO CHACABUCO
         !codigo_comuna %in% c(13202, 13203, 13302, 13303, 13402, 13403, 13404, 13601, 13602, 13603, 13605), #EXCLUYE COMUNAS FUERA DE LA RM
         !cod_zona %in% c(13401121001, 13124071001, 13124071002, 13124071004, 13124071005, 13124081001, 13124071003, 13119131001)) # QUITAR ZONAS FUERA DE LA MANCHA URBANA GS
# 13124071001 es "Ciudad de los Valles" (única zona fuera de la mancha urbana con medición ELSOC)

## JOIN: GEOMETRÍAS & ELSOC-ZONAS-ISMT
zonas_ams$cod_zona <- as.numeric(zonas_ams$cod_zona) #cambio clase para hacer el siguiente join
elsoc_ams_all <- dplyr::left_join(x = zonas_ams, y = elsoc_ams, by = "cod_zona") #TODAS LAS ZONAS AMS (SIRVE PARA MAPEAR)
  
# MAPA DE PRUEBA
ggplot(data = elsoc_ams_all) + geom_sf(aes(fill = ismt_rank, geometry = geometry)) + coord_sf(datum = st_crs(31979)) + scale_fill_viridis_c(direction = -1)

## CREATE SHAPEFILE
elsoc_ams_all_geo <- st_as_sf(x = elsoc_ams_all, sf_column_name = "geometry") %>% st_transform(31979)

```

```{r save-dataset, warning=FALSE}
# df para msem 
save(elsoc_ams, file = "../1_input/data/procesada/elsoc_ams.RData") #zonas elsoc
save(elsoc_ams_all, file = "../1_input/data/procesada/elsoc_ams_all.RData") #zonas ams

# df con coordenadas geográficas 
save(elsoc_ams_all_geo, file = "../1_input/data/procesada/elsoc_ams_all_geo.RData") 
st_write(elsoc_ams_all_geo, dsn = "../1_input/data/procesada/shapes/elsoc_ams_all_geo.shp", delete_layer = TRUE) 
```

```{r information}
sessionInfo()
```
