==> 2_processing/3-analisis-modelo-sem.Rmd <==
---
title: "Código de Análisis de Datos"
author: "Cristóbal Ortiz"
output: html_document
---

```{r Ajustes de Markdown,include=FALSE}
knitr::opts_chunk$set(warning = FALSE,message = FALSE, cache = TRUE)
knitr::opts_knit$set(root.dir = "../")
options(scipen=999)
options(knitr.kable.NA = '')
```

```{r Cargar paquetes, include=FALSE}
library(tidyverse)
library(lme4)
library(texreg)
library(vtable)
library(psych)
library(sjPlot)
library(cowplot)

library(lavaan)
library(lavaanPlot)
library(foreign)
library(sjmisc)
library(knitr)
```

```{r load-dataset-proc, include=FALSE}
getwd()
rm(list=ls())
load("../1_input/data/procesada/elsoc_barrio.RData")

#View(elsoc_sub)
```

```{r correlation}

elsoc_sub <- elsoc_barrio %>% 
  select(apbi, soci, segui, repbi, acci, hete) %>% 
  drop_na()

correl = cor(elsoc_sub)

pairs.panels(elsoc_sub,
             smooth = TRUE,      # If TRUE, draws loess smooths
             scale = FALSE,      # If TRUE, scales the correlation text font
             density = TRUE,     # If TRUE, adds density plots and histograms
             ellipses = F,       # If TRUE, draws ellipses
             method = "pearson", # Correlation method (also "spearman" or "kendall")
             pch = 21,           # pch symbol
             lm = TRUE,          # If TRUE, plots linear fit rather than the LOESS (smoothed) fit
             cor = TRUE,         # If TRUE, reports correlations
             jiggle = FALSE,     # If TRUE, data points are jittered
             factor = 2,         # Jittering factor
             hist.col = 4,       # Histograms color
             stars = TRUE,       # If TRUE, adds significance level with stars
             ci = TRUE)          # If TRUE, adds confidence intervals
```

```{r lrm-models}
#Path c': efecto de segregación sobre cohesión
fitmed1 <- lm(apbi ~ hete, data = elsoc_sub)
#Path b: efecto de reputación sobre cohesión
fitmed2 <- lm(apbi ~ hete + repbi, data = elsoc_sub)
#Path a: efecto de segregación sobre reputación
fitmed3 <- lm(repbi ~ hete, data = elsoc_sub)

summary(fitmed1) #path c': 0.04506 (no significativo)
summary(fitmed2) #path b: 0.33281***
summary(fitmed3) #path a: -0.3240** 
```

```{r sem-models}
med_1 <- ' # direct effect
             apbi ~ c*hete
           # mediator
             repbi ~ a*hete
             apbi ~ b*repbi
           # indirect effect (a*b)
             ab := a*b
           # total effect
             total := c + (a*b)
         '
fitmed4 <- sem(med_1, data = elsoc_sub, se = "bootstrap", bootstrap = 5000)

summary(fitmed4, standardized = TRUE) 

#path c'= 0.05 (no significativo)
#path b =  0.33***
#path a = -0.32** 
#a*b = -0.108**

#al ser c' significativo y a*b también, podemos hablar de una mediación parcial. En este caso, la heterogeneidad social del barrio ejerce un efecto negativo sobre el apego barrial, pero a través del aumento de la reputación residencial. Al revés, la homogeneidad social tiene un efecto positivo sobre el apego, sólo cuando disminuye la reputación.
```

```{r}
lavaanPlot(name = "model1", fitmed4, coefs = TRUE)
lavaanPlot(model = fitmed4, node_options = list(shape = "box", fontname = "Helvetica"), edge_options = list(color = "black"), coefs = TRUE, stars = "coefs")

lavaanExtra::nice_lavaanPlot(model = fitmed4, stand = F)
```
