==> 2_processing/3-analisis-msem_nse.Rmd <==
---
title: "Código de Análisis de Datos"
author: "Cristóbal Ortiz"
output: html_document
---

```{r setup-markdown,include=FALSE}
knitr::opts_chunk$set(warning = FALSE,message = FALSE, cache = TRUE)
knitr::opts_knit$set(root.dir = "../")
options(scipen=999)
options(knitr.kable.NA = '')
```

```{r load-library, include=FALSE}
library(tidyverse) #varios
library(sjmisc) #datos
library(texreg) #tablas
library(lavaan) #sem
library(semptools)
#devtools::install_github("dr-JT/semoutput") #de aquí sale sem_paths
library(semoutput)
```

```{r load-dataset-proc, include=FALSE}
getwd()
rm(list=ls())

load("../1_input/data/procesada/elsoc_ams.RData")
nse_data0 <- read.csv("../1_input/data/original/dterr-w01-w04/ismt_rank.csv") #NSE calculado con PCA por QRamond
#load("../3_output/modelos/sem_models.RData")

elsoc_sub0 <- elsoc_ams %>% 
  dplyr::select(idencuesta, cod_zona, 
                spbi, soci, 
                segui, repbi, sacci, desgr,
                edadi, educi, essui, time,
                ismt_rank) %>% 
  drop_na()

################################################
elsoc_ams_complete <- elsoc_ams %>% 
  dplyr::select(idencuesta, cod_zona) %>% 
  drop_na()
haven::write_dta(elsoc_ams_complete, "../1_input/data/procesada/elsoc_ams_complete.dta")
################################################

#view(dfSummary(elsoc_sub0, plain.ascii  = FALSE, style = "grid", graph.magnif = 0.75, valid.col = FALSE))

#nse_data <- nse_data0 %>% dplyr::select(idencuesta, ismt_rank)

#elsoc_sub <- dplyr::left_join(x = elsoc_sub0, y = nse_data, by = "idencuesta")

# Estandarización de variables
elsoc_sub <- elsoc_sub0 %>% mutate_at(vars(spbi, soci, segui, repbi, desgr, essui, edadi, educi, time, ismt_rank), scale)
```

```{r correlations}
elsoc_sub %>% 
  dplyr::select(spbi, soci, segui, repbi, sacci, desgr, edadi, educi, essui, time) %>% 
  psych::pairs.panels(smooth = TRUE,      # If TRUE, draws loess smooths
             scale = FALSE,      # If TRUE, scales the correlation text font
             density = TRUE,     # If TRUE, adds density plots and histograms
             ellipses = F,       # If TRUE, draws ellipses
             method = "pearson", # Correlation method (also "spearman" or "kendall")
             pch = 21,           # pch symbol
             lm = TRUE,          # If TRUE, plots linear fit rather than the LOESS (smoothed) fit
             cor = TRUE,         # If TRUE, reports correlations
             jiggle = FALSE,     # If TRUE, data points are jittered
             factor = 2,         # Jittering factor
             hist.col = 4,       # Histograms color
             stars = TRUE,       # If TRUE, adds significance level with stars
             ci = TRUE)          # If TRUE, adds confidence intervals

elsoc_sub %>% 
  dplyr::select(spbi, soci, ismt_rank) %>% 
  psych::pairs.panels(smooth = TRUE,      # If TRUE, draws loess smooths
             scale = FALSE,      # If TRUE, scales the correlation text font
             density = TRUE,     # If TRUE, adds density plots and histograms
             ellipses = F,       # If TRUE, draws ellipses
             method = "pearson", # Correlation method (also "spearman" or "kendall")
             pch = 21,           # pch symbol
             lm = TRUE,          # If TRUE, plots linear fit rather than the LOESS (smoothed) fit
             cor = TRUE,         # If TRUE, reports correlations
             jiggle = FALSE,     # If TRUE, data points are jittered
             factor = 2,         # Jittering factor
             hist.col = 4,       # Histograms color
             stars = TRUE,       # If TRUE, adds significance level with stars
             ci = TRUE)          # If TRUE, adds confidence intervals
```

# Modelos de ecuaciones estructurales (análisis de senderos multinivel)

## Mediación Multinivel: upper-level mediation (2-1-1)

```{r msem-model-spbi-soci, warning=FALSE}
msem_spbi <- '
level: 1
      # Modelo estructural nivel 1
      spbi ~ b1*repbi + b2*sacci + b3*segui  + b4*desgr + b5*edadi + b6*educi + b7*essui + b8*time
      
      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      educi ~~ essui
      edadi ~~ time
      
level: 2
      # Modelo estructural nivel 2
      spbi ~ b1*repbi + b2*sacci + b3*segui  + b4*desgr + c1*ismt_rank
      repbi ~ a1*ismt_rank
      sacci ~ a2*ismt_rank
      segui ~ a3*ismt_rank
      desgr ~ a4*ismt_rank

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      
      # Efecto indirecto 
      ind_ismt_repbi_spbi := a1*b1
      ind_ismt_sacci_spbi := a2*b2
      ind_ismt_segui_spbi := a3*b3
      ind_ismt_desgr_spbi := a4*b4

      # Efecto total 
      tot_ismt_repbi_spbi := c1 + (a1*b1)
      tot_ismt_sacci_spbi := c1 + (a2*b2)
      tot_ismt_segui_spbi := c1 + (a3*b3)
      tot_ismt_desgr_spbi := c1 + (a4*b4)
'

fit_msem_spbi <- sem(msem_spbi, data = elsoc_sub, cluster = "cod_zona", verbose = TRUE, optim.method = "em")
#summary(fit_msem_spbi, fit.measures = TRUE, standardized = TRUE)

msem_soci <- '
level: 1
      # Modelo estructural nivel 1
      soci ~ b1*repbi + b2*sacci + b3*segui  + b4*desgr + b5*edadi + b6*educi + b7*essui + b8*time

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      educi ~~ essui
      edadi ~~ time
      
level: 2
      # Modelo estructural nivel 2
      soci ~ b1*repbi + b2*sacci + b3*segui  + b4*desgr + c1*ismt_rank
      repbi ~ a1*ismt_rank
      sacci ~ a2*ismt_rank
      segui ~ a3*ismt_rank
      desgr ~ a4*ismt_rank

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      
      # Efecto indirecto 
      ind_ismt_repbi_soci := a1*b1
      ind_ismt_sacci_soci := a2*b2
      ind_ismt_segui_soci := a3*b3
      ind_ismt_desgr_soci := a4*b4

      # Efecto total 
      tot_ismt_repbi_soci := c1 + (a1*b1)
      tot_ismt_sacci_soci := c1 + (a2*b2)
      tot_ismt_segui_soci := c1 + (a3*b3)
      tot_ismt_desgr_soci := c1 + (a4*b4)
'

fit_msem_soci <- sem(msem_soci, data = elsoc_sub, cluster = "cod_zona", verbose = TRUE, optim.method = "em")
#summary(fit_msem_soci, fit.measures = TRUE, standardized = TRUE)
```

```{r coef-plot-2v, warning=FALSE}
# spbi
coef_spbi <-
  broom::tidy(fit_msem_spbi, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>% 
  filter(op %in% c("~", ":="), !str_detect(term, "^tot_"), !str_detect(label, "^a")) %>% 
  data_frame(var = c(rep("Pertenencia"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

# soci
coef_soci <-
  broom::tidy(fit_msem_soci, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>%    
  filter(op %in% c("~", ":="), !str_detect(term, "^tot_"), !str_detect(label, "^a")) %>% 
  data_frame(var = c(rep("Sociabilidad"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

coef_msem <- rbind(coef_spbi, coef_soci)

coef_msem <- coef_msem %>% 
  mutate(term = recode(term, 
                       "spbi ~ repbi" = "Reputación", "soci ~ repbi" = "Reputación",
                       "spbi ~ segui" = "Seguridad", "soci ~ segui" = "Seguridad",
                       "spbi ~ sacci" = "Satisfacción", "soci ~ sacci" = "Satisfacción",
                       "spbi ~ desgr" = "Conflicto", "soci ~ desgr" = "Conflicto",
                       
                       "spbi ~ ismt_rank" = "NSE barrio",  "soci ~ ismt_rank" = "NSE barrio",
                       "spbi ~ dens_pob" = "Densidad", "soci ~ dens_pob" = "Densidad",
                       
                       "ind_ismt_repbi_spbi := a1*b1" = "NSE barrio~\nReputación",
                       "ind_ismt_sacci_spbi := a2*b2" = "NSE barrio~\nSatisfacción",
                       "ind_ismt_segui_spbi := a3*b3" = "NSE barrio~\nSeguridad",
                       "ind_ismt_desgr_spbi := a4*b4" = "NSE barrio~\nConflicto",

                       "ind_ismt_repbi_soci := a1*b1" = "NSE barrio~\nReputación",
                       "ind_ismt_sacci_soci := a2*b2" = "NSE barrio~\nSatisfacción",
                       "ind_ismt_segui_soci := a3*b3" = "NSE barrio~\nSeguridad",
                       "ind_ismt_desgr_soci := a4*b4" = "NSE barrio~\nConflicto"))

coef_plot <- coef_msem %>% 
  ggplot(aes(x = factor(term, levels = c('NSE barrio', "Densidad", 
                                         'Reputación', 'Satisfacción', 'Seguridad', 'Conflicto',
                                         'NSE barrio~\nReputación', 'NSE barrio~\nSatisfacción',
                                         'NSE barrio~\nSeguridad', 'NSE barrio~\nConflicto')), 
           y = estimate)) +
  geom_hline(yintercept = 0, color = 'red') +
  geom_point(position = position_dodge(width = -0.6), size = 3) +
  geom_linerange(aes(ymin = conf.low, ymax = conf.high),
                 position = position_dodge(width = -0.6), 
                 size = .8) +
  coord_flip() +
  facet_wrap(.~ var) +
  theme(axis.text = element_text(size =  12),
        axis.title = element_text(size = 12, face = "bold"),
        strip.text = element_text(size = 12, face = "bold")) +
  labs(title = "",
       x = "",
       y = "Coeficientes estandarizados") +
  scale_x_discrete(limits = rev)

coef_plot

ggsave(coef_plot,filename = "../3_output/graficos/coef-nse-cob-iter.png", width = 8, height = 8, dpi = 300)
```

```{r msem-table-spbi-soci}
options(digits = 3)

#spbi
sem_table_spbi <- sem_paths(fit_msem_spbi, standardized = TRUE, ci = "standardized", ci_level = 0.95, print = TRUE)
#sem_table_spbi %>% cat(., file = "../3_output/tablas/sem_table_spbi.doc")
sem_table_spbi

#soci
sem_table_soci <- sem_paths(fit_msem_soci, standardized = TRUE, ci = "standardized", ci_level = 0.95, print = TRUE)
#sem_table_soci %>% cat(., file = "../3_output/tablas/sem_table_soci.doc")
sem_table_soci 
```

```{r fit-measure-spbi-soci}
#spbi
ajust_spbi <- fitMeasures(fit_msem_spbi, c("chisq", "cfi","rmsea","srmr")) %>% 
  as.data.frame() %>% 
  rownames_to_column("measure")

#writexl::write_xlsx(ajust_spbi, path ="../3_output/tablas/ajust_table_spbi.xlsx")

lavInspect(fit_msem_spbi, "icc")
lavInspect(fit_msem_spbi, "r2")

#soci
ajust_soci <- fitMeasures(fit_msem_soci, c("chisq", "cfi","rmsea","srmr")) %>% 
  as.data.frame() %>% 
  rownames_to_column("measure")

#writexl::write_xlsx(ajust_soci, path ="../3_output/tablas/ajust_table_soci.xlsx")

lavInspect(fit_msem_soci, "icc")
```

## Save models
```{r save-models}
#save(fit_msem_spbi, fit_msem_soci, file = "../3_output/modelos/sem_models.RData")
```
