==> 2_processing/3-analisis-msem_vit.Rmd <==
---
title: "Código de Análisis de Datos"
author: "Cristóbal Ortiz"
output: html_document
---

```{r setup-markdown,include=FALSE}
knitr::opts_chunk$set(warning = FALSE,message = FALSE, cache = TRUE)
knitr::opts_knit$set(root.dir = "../")
options(scipen=999)
options(knitr.kable.NA = '')
```

```{r load-library, include=FALSE}
library(tidyverse)
library(lme4)
library(texreg)
library(vtable)
library(psych)
library(sjPlot)
library(cowplot)

library(lavaan)
library(lavaanPlot)
library(semPlot)
library(foreign)
library(sjmisc)
library(knitr)
library(semTable)
library(semptools)
#devtools::install_github("dr-JT/semoutput") #de aquí sale sem_paths
library(semoutput)
library(stargazer)
```

```{r load-dataset-proc, include=FALSE}
getwd()
rm(list=ls())
load("../1_input/data/procesada/elsoc_ams.RData")
#View(elsoc_sub)

elsoc_sub <- elsoc_ams %>% 
  dplyr::select(idencuesta, cod_zona, 
                apbi, soci, 
                starts_with("t02_"), starts_with("t03_"),
                starts_with("t06_"),
                edadi, educi, essui, time,  
                ismt_rank, ismt_rank, gse_zona, theil_mixt, theil_segr, jane_indx) %>% 
  drop_na() %>% 
  rename(#pertenencia
         "spb1" = "t02_01_w06",
         "spb2" = "t02_02_w06",
         "spb3" = "t02_03_w06",
         "spb4" = "t02_04_w06",
         #sociabilidad
         "soc1" = "t03_01_w06",
         "soc2" = "t03_02_w06",
         "soc3" = "t03_03_w06",
         "soc4" = "t03_04_w06",
         #satisfacción
         "sat1" = "t06_01_w06",
         "sat2" = "t06_02_w06",
         "sat3" = "t06_03_w06",
         "sat4" = "t06_04_w06",
         "sat5" = "t06_05_w06",
         "sat6" = "t06_06_w06",
         "sat7" = "t06_07_w06",
         "sat8" = "t06_08_w06")
```

```{r correlations}
elsoc_sub %>% 
  dplyr::select(theil_pola, ismt_rank, theil_mixt, theil_segr, jane_indx) %>% 
  psych::pairs.panels(smooth = TRUE,      # If TRUE, draws loess smooths
             scale = FALSE,      # If TRUE, scales the correlation text font
             density = TRUE,     # If TRUE, adds density plots and histograms
             ellipses = F,       # If TRUE, draws ellipses
             method = "pearson", # Correlation method (also "spearman" or "kendall")
             pch = 21,           # pch symbol
             lm = TRUE,          # If TRUE, plots linear fit rather than the LOESS (smoothed) fit
             cor = TRUE,         # If TRUE, reports correlations
             jiggle = FALSE,     # If TRUE, data points are jittered
             factor = 2,         # Jittering factor
             hist.col = 4,       # Histograms color
             stars = TRUE,       # If TRUE, adds significance level with stars
             ci = TRUE)          # If TRUE, adds confidence intervals
```

# Modelos de ecuaciones estructurales (análisis de senderos multinivel)

## Análisis factorial confirmatorio (CFA)
```{r cfa-model}
sem1 <- '
# Latent variables
pertenencia =~ spb1 + spb2 + spb3 + spb4
sociabilidad =~ soc1 + soc2 + soc3 + soc4

satisfaccion =~ sat1 + sat2 + sat3 + sat4 + sat5 + sat6 + sat7 + sat8 

# Structural Model
pertenencia ~  satisfaccion 
sociabilidad ~ satisfaccion 

## Covarianzas
pertenencia ~~ sociabilidad
'

fitsem_1 <- sem(sem1, data = elsoc_sub, ordered = T)
#missing="fiml"
summary(fitsem_1, standardized = TRUE)
```

```{r cfa-table}
semTable(fitsem_1, type = "html",
         file = "resultados_cfa2",
         paramSets = c("loadings", "latentcovariances", "fits"))
```

```{r cfa-paths}

semPaths(fitsem_1, whatLabels = "est.std", intercepts = FALSE, nCharNodes = 0, edge.label.cex = 0.75, rotation = 2, residuals = F)

```
```{r}
my_position_list <- c("x4 ~ x1" = .75)
my_curve_list <- c("x2 ~ x1" = -2)
my_rotate_resid_list <- c(x1 = 0, x2 = 180, x3 = 140, x4 = 140)
my_position_list <- c("x4 ~ x1" = .65)
# If R version 4.1.0 or above
p_pa3 <- p_pa |> set_curve(my_curve_list) |>
                  rotate_resid(my_rotate_resid_list) |>
                  mark_sig(fit_pa) |>
                  mark_se(fit_pa, sep = "\n") |>
                  set_edge_label_position(my_position_list)
plot(p_pa3)
```


## Mediación Multinivel: upper-level mediation (2-1-1)
```{r msem-model-apbi-soci, warning=FALSE}
msem_apbi <- '
level: 1
      # Modelo estructural nivel 1
      apbi ~ b1*t06_01_w06 + b2*t06_02_w06 + b3*t06_03_w06 + b4*t06_04_w06 + b5*t06_05_w06 + b6*t06_06_w06 + b7*t06_07_w06 + b8*t06_08_w06 + b4*edadi + b5*educi + b6*essui + b7*time

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      educi ~~ essui
      edadi ~~ time
      
level: 2
      # Modelo estructural nivel 2
      apbi ~ t06_01_w06*b1 + t06_02_w06*b2 + t06_03_w06*b3 + t06_04_w06*b4 + t06_05_w06*b5 + t06_06_w06*b6 + t06_07_w06*b7 + t06_08_w06*b8 + c1*jane_conc + c2*jane_divu + c3*jane_edif + c4*jane_opco + c5*jane_acce + c6*jane_dist + c7*ismt_rank
      repbi ~ a1*ismt_rank
      segui ~ a2*ismt_rank
      sacci ~ a3*jane_indx
      segui ~ a4*jane_indx
      
      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      
      # Efecto indirecto 
      ind_segr_repbi_apbi := a1*b1
      ind_segr_segui_apbi := a2*b3
      ind_jane_sacci_apbi := a3*b2
      ind_jane_segui_apbi := a4*b3

      # Efecto total 
      tot_segr_repbi_apbi := c1 + (a1*b1)
      tot_segr_segui_apbi := c1 + (a2*b3)
      tot_jane_sacci_apbi := c2 + (a3*b2)
      tot_jane_segui_apbi := c2 + (a4*b3)
'

fit_msem_apbi <- sem(msem_apbi, data = elsoc_sub, cluster = "cod_zona", verbose = TRUE, fit.measures = TRUE, standardized = TRUE, optim.method = "em")
summary(fit_msem_apbi, fit.measures = T)

msem_soci <- '
level: 1
      # Modelo estructural nivel 1
      soci ~ t06_01_w06*b1 + t06_02_w06*b2 + t06_03_w06*b3 + t06_04_w06*b4 + t06_05_w06*b5 + t06_06_w06*b6 + t06_07_w06*b7 + t06_08_w06*b8 + b4*edadi + b5*educi + b6*essui + b7*time

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      educi ~~ essui
      edadi ~~ time
      
level: 2
      # Modelo estructural nivel 2
      soci ~ t06_01_w06*b1 + t06_02_w06*b2 + t06_03_w06*b3 + t06_04_w06*b4 + t06_05_w06*b5 + t06_06_w06*b6 + t06_07_w06*b7 + t06_08_w06*b8 + c1*jane_conc + c2*jane_divu + c3*jane_edif + c4*jane_opco + c5*jane_acce + c6*jane_dist + c7*ismt_rank
      repbi ~ a1*ismt_rank
      segui ~ a2*ismt_rank
      sacci ~ a3*jane_indx
      segui ~ a4*jane_indx
      
      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi

      # Efecto indirecto 
      ind_segr_repbi_soci := a1*b1
      ind_segr_segui_soci := a2*b3
      ind_jane_sacci_soci := a3*b2
      ind_jane_segui_soci := a4*b3

      # Efecto total 
      tot_segr_repbi_soci := c1 + (a1*b1)
      tot_segr_segui_soci := c1 + (a2*b3)
      tot_jane_sacci_soci := c2 + (a3*b2)
      tot_jane_segui_soci := c2 + (a4*b3)
'

fit_msem_soci <- sem(msem_soci, data = elsoc_sub, cluster = "cod_zona", verbose = TRUE, fit.measures = TRUE, standardized = TRUE, optim.method = "em")
#summary(fit_msem_soci, fit.measures = T)
```

```{r msem-table-apbi-soci}
options(digits = 3)

#apbi
sem_table_apbi <- sem_paths(fit_msem_apbi, standardized = TRUE, ci = "standardized", ci_level = 0.95, print = TRUE)
#sem_table_apbi %>% cat(., file = "../3_output/tablas/sem_table_apbi.doc")
sem_table_apbi

#soci
sem_table_soci <- sem_paths(fit_msem_soci, standardized = TRUE, ci = "standardized", ci_level = 0.95, print = TRUE)
#sem_table_soci %>% cat(., file = "../3_output/tablas/sem_table_soci.doc")
sem_table_soci 
```

```{r fit-measure-apbi-soci}
#apbi
ajust_apbi <- fitMeasures(fit_msem_apbi, c("chisq", "cfi","rmsea","srmr")) %>% 
  as.data.frame() %>% 
  rownames_to_column("measure")

writexl::write_xlsx(ajust_apbi, path ="../3_output/tablas/ajust_table_apbi.xlsx")

lavInspect(fit_msem_apbi, "icc")
lavInspect(fit_msem_apbi, "r2")

#soci
ajust_soci <- fitMeasures(fit_msem_soci, c("chisq", "cfi","rmsea","srmr")) %>% 
  as.data.frame() %>% 
  rownames_to_column("measure")

writexl::write_xlsx(ajust_soci, path ="../3_output/tablas/ajust_table_soci.xlsx")

lavInspect(fit_msem_soci, "icc")
```

```{r coef-plot, warning=FALSE}
# apbi
coef_apbi <-
  broom::tidy(fit_msem_apbi, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>% 
  filter(label %in% c("b1", "b2", "b3", "c1", "c2", "ind_segr_repbi_apbi", "ind_segr_segui_apbi", "ind_jane_sacci_apbi", "ind_jane_segui_apbi")) %>% 
  data_frame(var = c(rep("Pertenencia"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

# soci
coef_soci <-
  broom::tidy(fit_msem_soci, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>%    
  filter(label %in% c("b1", "b2", "b3", "c1", "c2", "ind_segr_repbi_soci", "ind_segr_segui_soci", "ind_jane_sacci_soci", "ind_jane_segui_soci")) %>% 
  data_frame(var = c(rep("Sociabilidad"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

coef_msem <- rbind(coef_apbi, coef_soci)

coef_msem <- coef_msem %>% 
  mutate(term = recode(term, 
                       "apbi ~ repbi" = "Reputación",
                       "soci ~ repbi" = "Reputación",
                       "apbi ~ segui" = "Seguridad",
                       "soci ~ segui" = "Seguridad",
                       "apbi ~ sacci" = "Satisfacción",
                       "soci ~ sacci" = "Satisfacción",
                       "apbi ~ ismt_rank" = "NSE barrio",
                       "soci ~ ismt_rank" = "NSE barrio",
                       "apbi ~ jane_indx" = "Vitalidad",
                       "soci ~ jane_indx" = "Vitalidad",
                       "ind_segr_repbi_apbi := a1*b1" = "NSE barrio ~\nReputación",
                       "ind_segr_repbi_soci := a1*b1" = "NSE barrio ~\nReputación",
                       "ind_segr_segui_apbi := a2*b3" = "NSE barrio ~\nSeguridad",
                       "ind_segr_segui_soci := a2*b3" = "NSE barrio ~\nSeguridad",
                       "ind_jane_sacci_apbi := a3*b2" = "Vitalidad ~\nSatisfacción",
                       "ind_jane_sacci_soci := a3*b2" = "Vitalidad ~\nSatisfacción",
                       "ind_jane_segui_apbi := a4*b3" = "Vitalidad ~\nSeguridad",
                       "ind_jane_segui_soci := a4*b3" = "Vitalidad ~\nSeguridad"))

coef_plot <- coef_msem %>% 
  ggplot(aes(x = factor(term, levels = c('Reputación', 'Satisfacción', 'Seguridad', "NSE barrio", "Vitalidad", 
                                         'NSE barrio ~\nReputación', 'NSE barrio ~\nSeguridad',
                                         'Vitalidad ~\nSatisfacción', 'Vitalidad ~\nSeguridad')), 
           y = estimate)) +
  geom_hline(yintercept = 0, color = 'red') +
  geom_point(position = position_dodge(width = -0.6), size = 3) +
  geom_linerange(aes(ymin = conf.low, ymax = conf.high),
                 position = position_dodge(width = -0.6), 
                 size = .8) +
  coord_flip() +
  facet_wrap(.~ var) +
  theme(axis.text = element_text(size =  12),
        axis.title = element_text(size = 12, face = "bold"),
        strip.text = element_text(size = 12, face = "bold")) +
  labs(title = "",
       x = "",
       y = "Coeficientes estandarizados") +
  scale_x_discrete(limits = rev)

coef_plot 

ggsave(coef_plot,filename = "../3_output/graficos/coef_plot_polar.png", width = 8, height = 6, dpi = 300)
```

## Save models
```{r save-models}
save(fit_msem_apbi, fit_msem_soci, file = "../3_output/modelos/sem_models.RData")

#save(fit_sem_apbi_low, fit_sem_apbi_middle, fit_sem_apbi_high, fit_sem_soci_low, fit_sem_soci_middle, fit_sem_soci_high, file = "../3_output/modelos/sem_models_gse.RData")
```
