==> 2_processing/3-analisis-sem-model_seg.Rmd <==
---
title: "Código de Análisis de Datos"
author: "Cristóbal Ortiz"
output: html_document
---

```{r setup-markdown,include=FALSE}
knitr::opts_chunk$set(warning = FALSE,message = FALSE, cache = TRUE)
knitr::opts_knit$set(root.dir = "../")
options(scipen=999)
options(knitr.kable.NA = '')
```

```{r load-library, include=FALSE}
library(tidyverse)
library(lme4)
library(texreg)
library(vtable)
library(psych)
library(sjPlot)
library(cowplot)

library(lavaan)
library(lavaanPlot)
library(semPlot)
library(foreign)
library(sjmisc)
library(knitr)
library(semTable)
library(semptools)
#devtools::install_github("dr-JT/semoutput") #de aquí sale sem_paths
library(semoutput)
library(stargazer)
```

```{r load-dataset-proc, include=FALSE}
getwd()
rm(list=ls())
load("../1_input/data/procesada/elsoc_ams.RData")
#load("../3_output/modelos/sem_models.RData")
#View(elsoc_sub)

elsoc_sub <- elsoc_ams %>% 
  dplyr::select(idencuesta, cod_zona, 
                apbi, soci, comi,
                segui, repbi, sacci, desgr,
                edadi, educi, essui, time,  
                starts_with("theil_"), sd_segr,
                starts_with("jane_"),
                ismt_rank, gse_zona) %>% 
  drop_na()

elsoc_apbi <- elsoc_ams %>% 
  dplyr::select(idencuesta, cod_zona, 
                apbi,
                segui, repbi, sacci, desgr,
                edadi, educi, essui, time,  
                starts_with("theil_"), sd_segr,
                starts_with("jane_"),
                ismt_rank, gse_zona) %>% 
  drop_na()

elsoc_soci <- elsoc_ams %>% 
  dplyr::select(idencuesta, cod_zona, 
                soci,
                segui, repbi, sacci, desgr,
                edadi, educi, essui, time,  
                starts_with("theil_"), sd_segr,
                starts_with("jane_"),
                ismt_rank, gse_zona) %>% 
  drop_na()

elsoc_comi <- elsoc_ams %>% 
  dplyr::select(idencuesta, cod_zona, 
                comi,
                segui, repbi, sacci, desgr,
                edadi, educi, essui, time,  
                starts_with("theil_"), sd_segr,
                starts_with("jane_"),
                ismt_rank, gse_zona) %>% 
  drop_na()
```

```{r correlations}
elsoc_sub %>% 
  dplyr::select(segui, repbi, sacci, desgr, edadi, educi, essui, time) %>% 
  psych::pairs.panels(smooth = TRUE,      # If TRUE, draws loess smooths
             scale = FALSE,      # If TRUE, scales the correlation text font
             density = TRUE,     # If TRUE, adds density plots and histograms
             ellipses = F,       # If TRUE, draws ellipses
             method = "pearson", # Correlation method (also "spearman" or "kendall")
             pch = 21,           # pch symbol
             lm = TRUE,          # If TRUE, plots linear fit rather than the LOESS (smoothed) fit
             cor = TRUE,         # If TRUE, reports correlations
             jiggle = FALSE,     # If TRUE, data points are jittered
             factor = 2,         # Jittering factor
             hist.col = 4,       # Histograms color
             stars = TRUE,       # If TRUE, adds significance level with stars
             ci = TRUE)          # If TRUE, adds confidence intervals

elsoc_sub %>% 
  dplyr::select(theil_pola, ismt_rank, sd_segr, theil_segr, jane_indx) %>% 
  psych::pairs.panels(smooth = TRUE,      # If TRUE, draws loess smooths
             scale = FALSE,      # If TRUE, scales the correlation text font
             density = TRUE,     # If TRUE, adds density plots and histograms
             ellipses = F,       # If TRUE, draws ellipses
             method = "pearson", # Correlation method (also "spearman" or "kendall")
             pch = 21,           # pch symbol
             lm = TRUE,          # If TRUE, plots linear fit rather than the LOESS (smoothed) fit
             cor = TRUE,         # If TRUE, reports correlations
             jiggle = FALSE,     # If TRUE, data points are jittered
             factor = 2,         # Jittering factor
             hist.col = 4,       # Histograms color
             stars = TRUE,       # If TRUE, adds significance level with stars
             ci = TRUE)          # If TRUE, adds confidence intervals
```

# Modelos de ecuaciones estructurales (análisis de senderos multinivel)

## Mediación Multinivel: upper-level mediation (2-1-1)

```{r msem-model-apbi-soci, warning=FALSE}
msem_apbi <- '
level: 1
      # Modelo estructural nivel 1
      apbi ~ b1*repbi + b2*segui + b3*sacci + b4*desgr + b5*edadi + b6*educi + b7*essui + b8*time

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      educi ~~ essui
      edadi ~~ time
      
level: 2
      # Modelo estructural nivel 2
      apbi ~ b1*repbi + b2*segui + b3*sacci + b4*desgr + c1*theil_pola + c2*sd_segr + c3*jane_indx
      repbi ~ a1*theil_pola
      segui ~ a2*theil_pola
      sacci ~ a3*theil_pola
      desgr ~ a4*theil_pola

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      
      # Efecto indirecto 
      ind_segr_repbi_apbi := a1*b1
      ind_segr_segui_apbi := a2*b2
      ind_segr_sacci_apbi := a3*b3
      ind_segr_desgr_apbi := a4*b4

      # Efecto total 
      tot_segr_repbi_apbi := c1 + (a1*b1)
      tot_segr_segui_apbi := c1 + (a2*b2)
      tot_segr_sacci_apbi := c1 + (a3*b3)
      tot_segr_desgr_apbi := c1 + (a4*b4)
'

fit_msem_apbi <- sem(msem_apbi, data = elsoc_apbi, cluster = "cod_zona", verbose = TRUE, optim.method = "em")
summary(fit_msem_apbi, fit.measures = T)

msem_soci <- '
level: 1
      # Modelo estructural nivel 1
      soci ~ b1*repbi + b2*segui + b3*sacci + b4*desgr + b5*edadi + b6*educi + b7*essui + b8*time

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      educi ~~ essui
      edadi ~~ time
      
level: 2
      # Modelo estructural nivel 2
      soci ~ b1*repbi + b2*segui + b3*sacci + b4*desgr + c1*theil_pola + c2*sd_segr + c3*jane_indx
      repbi ~ a1*theil_pola
      segui ~ a2*theil_pola
      sacci ~ a3*theil_pola
      desgr ~ a4*theil_pola

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      
      # Efecto indirecto 
      ind_segr_repbi_soci := a1*b1
      ind_segr_segui_soci := a2*b2
      ind_segr_sacci_soci := a3*b3
      ind_segr_desgr_soci := a4*b4

      # Efecto total 
      tot_segr_repbi_soci := c1 + (a1*b1)
      tot_segr_segui_soci := c1 + (a2*b2)
      tot_segr_sacci_soci := c1 + (a3*b3)
      tot_segr_desgr_soci := c1 + (a4*b4)
'

fit_msem_soci <- sem(msem_soci, data = elsoc_soci, cluster = "cod_zona", verbose = TRUE, optim.method = "em")
#summary(fit_msem_soci, fit.measures = T)


msem_comi <- '
level: 1
      # Modelo estructural nivel 1
      comi ~ b1*repbi + b2*segui + b3*sacci + b4*desgr + b5*edadi + b6*educi + b7*essui + b8*time

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      educi ~~ essui
      edadi ~~ time
      
level: 2
      # Modelo estructural nivel 2
      comi ~ b1*repbi + b2*segui + b3*sacci + b4*desgr + c1*theil_pola + c2*sd_segr + c3*jane_indx
      repbi ~ a1*theil_pola
      segui ~ a2*theil_pola
      sacci ~ a3*theil_pola
      desgr ~ a4*theil_pola

      # Covarianzas
      repbi ~~ segui
      segui ~~ sacci
      sacci ~~ repbi
      desgr ~~ repbi
      desgr ~~ sacci
      desgr ~~ segui
      
      # Efecto indirecto 
      ind_segr_repbi_soci := a1*b1
      ind_segr_segui_soci := a2*b2
      ind_segr_sacci_soci := a3*b3
      ind_segr_desgr_soci := a4*b4

      # Efecto total 
      tot_segr_repbi_comi := c1 + (a1*b1)
      tot_segr_segui_comi := c1 + (a2*b2)
      tot_segr_sacci_comi := c1 + (a3*b3)
      tot_segr_desgr_comi := c1 + (a4*b4)
'

fit_msem_comi <- sem(msem_comi, data = elsoc_comi, cluster = "cod_zona", verbose = TRUE, optim.method = "em")
#summary(fit_msem_comi, fit.measures = T)
```

```{r msem-table-apbi-soci}
options(digits = 3)

#apbi
sem_table_apbi <- sem_paths(fit_msem_apbi, standardized = TRUE, ci = "standardized", ci_level = 0.95, print = TRUE)
#sem_table_apbi %>% cat(., file = "../3_output/tablas/sem_table_apbi.doc")
sem_table_apbi

#soci
sem_table_soci <- sem_paths(fit_msem_soci, standardized = TRUE, ci = "standardized", ci_level = 0.95, print = TRUE)
#sem_table_soci %>% cat(., file = "../3_output/tablas/sem_table_soci.doc")
sem_table_soci 
```

```{r fit-measure-apbi-soci}
#apbi
ajust_apbi <- fitMeasures(fit_msem_apbi, c("chisq", "cfi","rmsea","srmr")) %>% 
  as.data.frame() %>% 
  rownames_to_column("measure")

#writexl::write_xlsx(ajust_apbi, path ="../3_output/tablas/ajust_table_apbi.xlsx")

lavInspect(fit_msem_apbi, "icc")
lavInspect(fit_msem_apbi, "r2")

#soci
ajust_soci <- fitMeasures(fit_msem_soci, c("chisq", "cfi","rmsea","srmr")) %>% 
  as.data.frame() %>% 
  rownames_to_column("measure")

#writexl::write_xlsx(ajust_soci, path ="../3_output/tablas/ajust_table_soci.xlsx")

lavInspect(fit_msem_soci, "icc")
```

```{r coef-plot-2v, warning=FALSE}
# apbi
coef_apbi <-
  broom::tidy(fit_msem_apbi, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>% 
  filter(op %in% c("~", ":="), !str_detect(term, "^tot_"), !str_detect(label, "^a")) %>% 
  data_frame(var = c(rep("Pertenencia"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

# soci
coef_soci <-
  broom::tidy(fit_msem_soci, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>%    
  filter(op %in% c("~", ":="), !str_detect(term, "^tot_"), !str_detect(label, "^a")) %>% 
  data_frame(var = c(rep("Sociabilidad"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

coef_msem <- rbind(coef_apbi, coef_soci)

coef_msem <- coef_msem %>% 
  mutate(term = recode(term, 
                       "apbi ~ repbi" = "Reputación",
                       "soci ~ repbi" = "Reputación",
                       "apbi ~ segui" = "Seguridad",
                       "soci ~ segui" = "Seguridad",
                       "apbi ~ sacci" = "Satisfacción",
                       "soci ~ sacci" = "Satisfacción",
                       "apbi ~ desgr" = "Conflicto",
                       "soci ~ desgr" = "Conflicto",
                       "apbi ~ theil_pola" = "Segregación",
                       "soci ~ theil_pola" = "Segregación",
                       "apbi ~ sd_segr" = "Homogeneidad",
                       "soci ~ sd_segr" = "Homogeneidad",
                       "apbi ~ jane_indx" = "Vitalidad",
                       "soci ~ jane_indx" = "Vitalidad",
                       "ind_segr_repbi_apbi := a1*b1" = "Segregación~\nReputación",
                       "ind_segr_segui_apbi := a2*b2" = "Segregación~\nSeguridad",
                       "ind_segr_sacci_apbi := a3*b3" = "Segregación~\nSatisfacción",
                       "ind_segr_desgr_apbi := a4*b4" = "Segregación~\nConflicto",
                       "ind_segr_desga_apbi := a4*b5" = "Segregación~\nActividades",
                       "ind_segr_repbi_soci := a1*b1" = "Segregación~\nReputación",
                       "ind_segr_segui_soci := a2*b2" = "Segregación~\nSeguridad",
                       "ind_segr_sacci_soci := a3*b3" = "Segregación~\nSatisfacción",
                       "ind_segr_desgr_soci := a4*b4" = "Segregación~\nConflicto",
                       "ind_segr_desga_soci := a4*b5" = "Segregación~\nActividades"))

coef_plot <- coef_msem %>% 
  ggplot(aes(x = factor(term, levels = c('Reputación', 'Satisfacción', 'Seguridad', 'Conflicto',
                                         "Segregación", 'Homogeneidad', "Vitalidad", 
                                         'Segregación~\nReputación', 'Segregación~\nSeguridad',
                                         'Segregación~\nSatisfacción', 'Segregación~\nConflicto')), 
           y = estimate)) +
  geom_hline(yintercept = 0, color = 'red') +
  geom_point(position = position_dodge(width = -0.6), size = 3) +
  geom_linerange(aes(ymin = conf.low, ymax = conf.high),
                 position = position_dodge(width = -0.6), 
                 size = .8) +
  coord_flip() +
  facet_wrap(.~ var) +
  theme(axis.text = element_text(size =  12),
        axis.title = element_text(size = 12, face = "bold"),
        strip.text = element_text(size = 12, face = "bold")) +
  labs(title = "",
       x = "",
       y = "Coeficientes estandarizados") +
  scale_x_discrete(limits = rev)

coef_plot 

ggsave(coef_plot,filename = "../3_output/graficos/coef_plot_seg_cob.png", width = 8, height = 6, dpi = 300)
```

```{r coef-plot-3v, warning=FALSE}
# apbi
coef_apbi <-
  broom::tidy(fit_msem_apbi, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>% 
  filter(op %in% c("~", ":="), !str_detect(term, "^tot_"), !str_detect(label, "^a")) %>% 
  data_frame(var = c(rep("Pertenencia"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

# soci
coef_soci <-
  broom::tidy(fit_msem_soci, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>%    
  filter(op %in% c("~", ":="), !str_detect(term, "^tot_"), !str_detect(label, "^a")) %>% 
  data_frame(var = c(rep("Sociabilidad"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

coef_comi <-
  broom::tidy(fit_msem_comi, conf.int = TRUE, conf.level = 0.95) %>%
  filter(!level == 1) %>%    
  filter(op %in% c("~", ":="), !str_detect(term, "^tot_"), !str_detect(label, "^a")) %>% 
  data_frame(var = c(rep("Compromiso"))) %>% 
  dplyr::select(var, term, estimate, conf.low, conf.high)

coef_msem <- rbind(coef_apbi, coef_soci, coef_comi)

coef_msem <- coef_msem %>% 
  mutate(term = recode(term, 
                       "apbi ~ repbi" = "Reputación",
                       "soci ~ repbi" = "Reputación",
                       "comi ~ repbi" = "Reputación",
                       "apbi ~ segui" = "Seguridad",
                       "soci ~ segui" = "Seguridad",
                       "comi ~ segui" = "Seguridad",
                       "apbi ~ sacci" = "Satisfacción",
                       "soci ~ sacci" = "Satisfacción",
                       "comi ~ sacci" = "Satisfacción",
                       "apbi ~ desgr" = "Conflicto",
                       "soci ~ desgr" = "Conflicto",
                       "comi ~ desgr" = "Conflicto",
                       "apbi ~ theil_pola" = "Segregación",
                       "soci ~ theil_pola" = "Segregación",
                       "comi ~ theil_pola" = "Segregación",
                       "apbi ~ sd_segr" = "Homogeneidad",
                       "soci ~ sd_segr" = "Homogeneidad",
                       "comi ~ sd_segr" = "Homogeneidad",
                       "apbi ~ jane_indx" = "Vitalidad",
                       "soci ~ jane_indx" = "Vitalidad",
                       "comi ~ jane_indx" = "Vitalidad",
                       "ind_segr_repbi_apbi := a1*b1" = "Segregación~\nReputación",
                       "ind_segr_segui_apbi := a2*b2" = "Segregación~\nSeguridad",
                       "ind_segr_sacci_apbi := a3*b3" = "Segregación~\nSatisfacción",
                       "ind_segr_desgr_apbi := a4*b4" = "Segregación~\nConflicto",
                       "ind_segr_desga_apbi := a4*b5" = "Segregación~\nActividades",
                       "ind_segr_repbi_soci := a1*b1" = "Segregación~\nReputación",
                       "ind_segr_segui_soci := a2*b2" = "Segregación~\nSeguridad",
                       "ind_segr_sacci_soci := a3*b3" = "Segregación~\nSatisfacción",
                       "ind_segr_desgr_soci := a4*b4" = "Segregación~\nConflicto",
                       "ind_segr_desga_soci := a4*b5" = "Segregación~\nActividades",
                       "ind_segr_repbi_comi := a1*b1" = "Segregación~\nReputación",
                       "ind_segr_segui_comi := a2*b2" = "Segregación~\nSeguridad",
                       "ind_segr_sacci_comi := a3*b3" = "Segregación~\nSatisfacción",
                       "ind_segr_desgr_comi := a4*b4" = "Segregación~\nConflicto",
                       "ind_segr_desga_comi := a4*b5" = "Segregación~\nActividades"))

coef_plot <- coef_msem %>% 
  ggplot(aes(x = factor(term, levels = c('Reputación', 'Satisfacción', 'Seguridad', 'Conflicto',
                                         "Segregación", 'Homogeneidad', "Vitalidad", 
                                         'Segregación~\nReputación', 'Segregación~\nSeguridad',
                                         'Segregación~\nSatisfacción', 'Segregación~\nConflicto')), 
           y = estimate)) +
  geom_hline(yintercept = 0, color = 'red') +
  geom_point(position = position_dodge(width = -0.6), size = 3) +
  geom_linerange(aes(ymin = conf.low, ymax = conf.high),
                 position = position_dodge(width = -0.6), 
                 size = .8) +
  coord_flip() +
  facet_wrap(.~ var) +
  theme(axis.text = element_text(size =  12),
        axis.title = element_text(size = 12, face = "bold"),
        strip.text = element_text(size = 12, face = "bold")) +
  labs(title = "",
       x = "",
       y = "Coeficientes estandarizados") +
  scale_x_discrete(limits = rev)

coef_plot 

#ggsave(coef_plot,filename = "../3_output/graficos/coef_plot_seg_cob.png", width = 8, height = 6, dpi = 300)
```


## Save models
```{r save-models}
save(fit_msem_apbi, fit_msem_soci, file = "../3_output/modelos/sem_models.RData")

#save(fit_sem_apbi_low, fit_sem_apbi_middle, fit_sem_apbi_high, fit_sem_soci_low, fit_sem_soci_middle, fit_sem_soci_high, file = "../3_output/modelos/sem_models_gse.RData")
```
